feat: Cell UIView로 분리 (#4)
- 반복되는 디자인의 뷰를 따로 분리해서 관리
- 모든 셀에서 해당 셀 UIView를 사용함으로 재사용성을 용이하게 수정
